Remove unused script tags and update npm scripts
This commit removes unused script tags in the index.html file to
declutter the code and improve readability. Additionally, the
package.json file is updated to remove unused npm scripts and
dependencies, such as "devDependencies" and "nodemon". The "test",
"start", "update", and "format" scripts are also simplified and replaced
with "u" and "f" for git pull and npm i, and running Prettier
respectively. These changes optimize the codebase by removing
unnecessary code and improving the development workflow.

Issue: N/A